Set env-aware feature fallbacks
